// src/common/enums/property-type.enum.ts
// src/common/enums/property-type.enum.ts
export enum PropertyType {
  APARTMENT = 'apartment', // Квартира
  HOUSE = 'house', // Частный дом
  COMMERCIAL = 'commercial', // Коммерческая недвижимость
  LAND = 'land', // Земельный участок
  GARAGE = 'garage', // Гараж
  OFFICE = 'office', // Офис
  WAREHOUSE = 'warehouse', // Склад
  COTTAGE = 'cottage', // Коттедж
}