// src/common/enums/property-status.enum.ts
// src/common/enums/property-status.enum.ts
export enum PropertyStatus {
  ACTIVE = 'active', // Активный
  INACTIVE = 'inactive', // Неактивный
  SOLD = 'sold', // Продан
  RENTED = 'rented', // Сдан в аренду
  RESERVED = 'reserved', // Забронирован
  UNDER_CONSTRUCTION = 'under_construction', // На стадии строительства
}