// src/common/enums/property-tag.enum.ts
// src/common/enums/property-tag.enum.ts
export enum PropertyTag {
  HOT = 'hot', // Горячее предложение
  TOP = 'top', // Топ
  NEW = 'new', // Новое
  PROMOTION = 'promotion', // Акция
  IN_MORTGAGE = 'in_mortgage', // В залоге
  URGENT = 'urgent', // Срочно
  LUXURY = 'luxury', // Люкс
}