// src/common/enums/amenity.enum.ts
// src/common/enums/amenity.enum.ts
export enum Amenity {
  // Основные удобства
  INTERNET = 'internet', // Интернет
  PHONE = 'phone', // Телефон
  ELECTRICITY = 'electricity', // Электричество
  WATER = 'water', // Вода
  GAS = 'gas', // Газ
  HEATING = 'heating', // Отопление
  AIR_CONDITIONING = 'air_conditioning', // Кондиционер
  SEWERAGE = 'sewerage', // Канализация

  // Кухня
  KITCHEN = 'kitchen', // Кухня
  REFRIGERATOR = 'refrigerator', // Холодильник
  STOVE = 'stove', // Плита
  OVEN = 'oven', // Духовка
  MICROWAVE = 'microwave', // Микроволновка
  DISHWASHER = 'dishwasher', // Посудомоечная машина
  KETTLE = 'kettle', // Чайник

  // Ванная комната
  BATHROOM = 'bathroom', // Ванная комната
  SHOWER = 'shower', // Душ
  BATHTUB = 'bathtub', // Ванна
  WASHING_MACHINE = 'washing_machine', // Стиральная машина
  DRYER = 'dryer', // Сушилка

  // Мебель
  FURNISHED = 'furnished', // Меблирована
  BED = 'bed', // Кровать
  SOFA = 'sofa', // Диван
  TABLE = 'table', // Стол
  CHAIRS = 'chairs', // Стулья
  WARDROBE = 'wardrobe', // Шкаф

  // Безопасность
  SECURITY = 'security', // Охрана
  INTERCOM = 'intercom', // Домофон
  DOOR_PHONE = 'door_phone', // Видеодомофон
  CCTV = 'cctv', // Видеонаблюдение
  FIRE_ALARM = 'fire_alarm', // Пожарная сигнализация

  // Парковка
  PARKING = 'parking', // Парковка
  GARAGE = 'garage', // Гараж
  UNDERGROUND_PARKING = 'underground_parking', // Подземная парковка

  // Балкон/лоджия
  BALCONY = 'balcony', // Балкон
  LOGGIA = 'loggia', // Лоджия
  TERRACE = 'terrace', // Терраса

  // Дополнительно
  ELEVATOR = 'elevator', // Лифт
  CONCIERGE = 'concierge', // Консьерж
  GARDEN = 'garden', // Сад
  PLAYGROUND = 'playground', // Детская площадка
  GYM = 'gym', // Спортзал
  POOL = 'pool', // Бассейн
  SAUNA = 'sauna', // Сауна
  SPA = 'spa', // Спа
  STORAGE = 'storage', // Кладовая
  WALK_IN_CLOSET = 'walk_in_closet', // Гардеробная
  FIREPLACE = 'fireplace', // Камин
  PETS_ALLOWED = 'pets_allowed', // Можно с животными
  FLOOR_HEATING = 'floor_heating', // Теплый пол
  CEILING_FAN = 'ceiling_fan', // Потолочный вентилятор
  SMART_HOME = 'smart_home', // Умный дом
}